--- a/cookbooks/bcpc/recipes/ceph-head.rb
+++ b/cookbooks/bcpc/recipes/ceph-head.rb
@@ -213,7 +213,7 @@ if get_head_nodes.length == 1; then
     end
 end
 
-replicas = [search_nodes("recipe", "ceph-work").length, node['bcpc']['ceph']['default']['replicas']].min
+replicas = [search_nodes("recipe", "ceph-osd").length, node['bcpc']['ceph']['default']['replicas']].min
 if replicas < 1; then
     replicas = 1
 end