--- a/cookbooks/bcpc/recipes/ceph-rgw.rb
+++ b/cookbooks/bcpc/recipes/ceph-rgw.rb
@@ -74,7 +74,7 @@ rgw_rule = (node['bcpc']['ceph']['rgw']['type'] == "ssd") ? node['bcpc']['ceph']
     end
     bash "set-#{pool}-rados-pool-replicas" do
         user "root"
-        replicas = [search_nodes("recipe", "ceph-work").length, node['bcpc']['ceph']['rgw']['replicas']].min
+        replicas = [search_nodes("recipe", "ceph-osd").length, node['bcpc']['ceph']['rgw']['replicas']].min
         if replicas < 1; then
             replicas = 1
         end