--- a/cookbooks/bcpc/recipes/ceph-work.rb
+++ b/cookbooks/bcpc/recipes/ceph-work.rb
@@ -41,4 +41,3 @@ bash "write-bootstrap-osd-key" do
     EOH
     not_if "test -f /var/lib/ceph/bootstrap-osd/ceph.keyring"
 end
-