--- a/cookbooks/bcpc/recipes/cinder.rb
+++ b/cookbooks/bcpc/recipes/cinder.rb
@@ -105,7 +105,7 @@ node['bcpc']['ceph']['enabled_pools'].each do |type|
 
     bash "set-cinder-rados-pool-replicas-#{type}" do
         user "root"
-        replicas = [search_nodes("recipe", "ceph-work").length, node['bcpc']['ceph']['volumes']['replicas']].min
+        replicas = [search_nodes("recipe", "ceph-osd").length, node['bcpc']['ceph']['volumes']['replicas']].min
         if replicas < 1; then
             replicas = 1
         end