--- a/roles/BCPC-EphemeralWorknode.json
+++ b/roles/BCPC-EphemeralWorknode.json
@@ -1,0 +1,20 @@
+{
+    "name": "BCPC-EphemeralWorknode",
+    "default_attributes": {
+    },
+    "json_class": "Chef::Role",
+    "run_list": [
+      "role[BCPC-Ephemeral]"
+    ],
+    "description": "A functional compute node in a BCPC cluster",
+    "chef_type": "role",
+    "override_attributes": {
+      "bcpc" : {
+          "aggregate_membership" : ["ephemeral_compute"],
+	  "nova" : {
+	      "ephemeral" : true,
+	      "ephemeral_disks": ["/dev/sdb", "/dev/sdc"]
+	  }
+      }
+    }
+}