Basically a tidied up version of #687, but with one important difference. The volumes are directly creating in LVM. The previous PR actually used LVM to create a file system then put a file system over that. The LVM drive is considerably more performant (from what I have read).

I also explicitly create the flavors using the environment and the osflavor LWRP and give the ephemeral disk extra scratch space. 

Future interesting features:
* I believe that you can tell nova to format the ephemeral block dev, so it would appear mounted and available as scratch. This is probably worth doing, or at least finding out if cloud-init can be configured to do it. 
* Cinder often needs a kick after the head node run, I think this is an issue with cephx auth and will probably be fixed by #720.