@cloudm2 2015-07-25
I know they claim it but I find it hard to believe that nova lvm is more performant. Nothing in openstack is more performant than native.