@caiush 2015-07-27
the LVM driver doesn't have the overhead of the file system, thats all.  It has to be faster, if its less stable.. TBD... but its pretty mature, been in openstack since E for F (don't recall from  memory).