@cloudm2 2015-07-27
You still have the normal pvcreate and vgcreate. The only thing missing in the recipe is lvcreate which I would believe the nova "driver" (not a real driver I'm sure - haven't looked into the code) will call since nothing in Openstack is really performant. Not throwing a wet blanket but I don't understand why spent more cycles on this when not really needed.