@erhudy 2015-07-30
On a 6-node local build (1 head node, 1 RBD work node, 1 ephemeral work node, 3 monitoring nodes), Ceph looks pretty unhappy:
```
    cluster bb93c222-d3e4-4643-8f09-8396e6e9c28e
     health HEALTH_WARN
            2290 pgs degraded
            2290 pgs stuck degraded
            704 pgs stuck unclean
            2290 pgs stuck undersized
            2290 pgs undersized
            recovery 112/183 objects degraded (61.202%)
            recovery 7/183 objects misplaced (3.825%)
            too many PGs per OSD (624 > max 300)
     monmap e1: 1 mons at {bcpc-vm1=172.16.100.11:6789/0}
            election epoch 2, quorum 0 bcpc-vm1
     osdmap e166: 8 osds: 8 up, 8 in; 206 remapped pgs
      pgmap v4654: 2496 pgs, 18 pools, 40164 kB data, 61 objects
            454 MB used, 143 GB / 143 GB avail
            112/183 objects degraded (61.202%)
            7/183 objects misplaced (3.825%)
                1792 active+clean+undersized+degraded
                 498 active+undersized+degraded
                 206 active+remapped
```