@erhudy 2015-07-30
Suffered a failure when cheffing the **BCPC-EphemeralWorknode** role onto a physical node with 7 available SSDs:
```
Recipe: bcpc::lvm
  * apt_package[lvm2] action install
    - install version 2.02.98-6ubuntu2 of package lvm2
  * bash[setup-lvm-pv] action run
    
    ================================================================================
    Error executing action `run` on resource 'bash[setup-lvm-pv]'
    ================================================================================
    
    Mixlib::ShellOut::ShellCommandFailed
    ------------------------------------
    Expected process to exit with [0], but received '5'
    ---- Begin output of "bash"  "/tmp/chef-script20150730-2000-18only2" ----
    STDOUT: 
    STDERR: Device /dev/sdb not found (or ignored by filtering).
      Device /dev/sdc not found (or ignored by filtering).
      Device /dev/sdd not found (or ignored by filtering).
      Device /dev/sde not found (or ignored by filtering).
      Device /dev/sdf not found (or ignored by filtering).
      Device /dev/sdg not found (or ignored by filtering).
      Device /dev/sdh not found (or ignored by filtering).
    ---- End output of "bash"  "/tmp/chef-script20150730-2000-18only2" ----
    Ran "bash"  "/tmp/chef-script20150730-2000-18only2" returned 5
    
    Resource Declaration:
    ---------------------
    # In /var/chef/cache/cookbooks/bcpc/recipes/lvm.rb
    
     23:   bash "setup-lvm-pv" do
     24:     user "root"
     25:     code <<-EOH
     26:     pvcreate #{ node['bcpc']['nova']['ephemeral_disks'].join(' ') }
     27:   EOH
     28:     not_if "pvdisplay | grep '/dev'"
     29:   end
     30: 
    
    Compiled Resource:
    ------------------
    # Declared in /var/chef/cache/cookbooks/bcpc/recipes/lvm.rb:23:in `from_file'
    
    bash("setup-lvm-pv") do
      action "run"
      retries 0
      retry_delay 2
      default_guard_interpreter :default
      command "setup-lvm-pv"
      backup 5
      returns 0
      user "root"
      code "    pvcreate /dev/sdb /dev/sdc /dev/sdd /dev/sde /dev/sdf /dev/sdg /dev/sdh\n"
      interpreter "bash"
      declared_type :bash
      cookbook_name "bcpc"
      recipe_name "lvm"
      not_if "pvdisplay | grep '/dev'"
    end
```