@erhudy 2015-07-30
The disks are available:
```
# ls -l /dev/sd*
brw-rw---- 1 root disk 8,   0 Jul 30 09:36 /dev/sda
brw-rw---- 1 root disk 8,   1 Jul 30 09:36 /dev/sda1
brw-rw---- 1 root disk 8,   2 Jul 30 09:36 /dev/sda2
brw-rw---- 1 root disk 8,  16 Jul 30 09:55 /dev/sdb
brw-rw---- 1 root disk 8,  32 Jul 30 09:55 /dev/sdc
brw-rw---- 1 root disk 8,  48 Jul 30 09:55 /dev/sdd
brw-rw---- 1 root disk 8,  64 Jul 30 09:55 /dev/sde
brw-rw---- 1 root disk 8,  80 Jul 30 09:55 /dev/sdf
brw-rw---- 1 root disk 8,  96 Jul 30 09:55 /dev/sdg
brw-rw---- 1 root disk 8, 112 Jul 30 09:55 /dev/sdh
```