@erhudy 2015-07-30
Needed to completely destroy GPT/MBR partition tables on the disks with gdisk (install gdisk, then for each disk, do `gdisk /dev/sdx`, **x**, **z**, answer Y to both questions). Not sure if gdisk is scriptable, but it's faster than zeroing the entire disk to destroy all copies of the GPT.