@erhudy 2015-07-30
Better way: `sgdisk -Z /dev/sdx`