@erhudy 2015-07-31
Re-opening so that I can continue to test and comment here. I think some of the issues that were being had previously were because of conflicts with some other unfinished code.