@erhudy 2015-07-31
Behavioral note: if you boot from image with an ephemeral flavor, both disks will be on LVM. If you boot from volume, you will get your root volume from Ceph and the ephemeral volume from LVM.