@erhudy 2015-08-01
Rebased branch off current master and pushed a fix for calculating Ceph PGs, so that the recipe is only counting nodes that actually have OSDs on them.