@erhudy 2015-08-03
Here are some informal results of quick and dirty testing on hardware (note that this is only using 21 OSDs, 7 per system, because of re-dedicating some hardware to ephemeral).

Test conditions
---
* e1.medium running Ubuntu 14.04.2 LTS with Ceph root partition on `/dev/vda` and LVM ephemeral partition on `/dev/vdb` (same as m1.medium, plus ephemeral disk)
* formatted `/dev/vdb` as ext4
* rest of cluster was quiesced to minimize traffic from anywhere else (so no noisy neighbor testing)
* readahead set to 2048KB on all block devices via `blockdev --setra`
* using `bonnie++`, ran 5 times with default settings as `ubuntu` (8GB test payload)

Results
---
* ephemeral is faster for reads: ~500MB/sec for ephemeral versus ~400MB/sec for Ceph, though Ceph has turned in slightly better results with larger Ceph installations
* Ceph is faster for block writes by a considerable margin: ~200MB/sec for ephemeral versus ~300MB/sec for Ceph, and Ceph again has turned in better results with larger installations
* most curiously, Ceph completely clowns ephemeral in terms of latency, which is counter-intuitive and is leading me to doubt the test methodology somewhat
  * ~125ms block writes for Ceph versus ~300ms for ephemeral
  * ~250ms block rewrites for Ceph versus ~850ms for ephemeral
  * ~9ms block reads for Ceph versus ~14ms for ephemeral
* re-testing ephemeral with the smaller default block size of 256KB makes it considerably more performant for both reads and writes, though with unusual effects on latency
  * from ~200MB/sec to ~260MB/sec for block writes
  * from ~500MB/sec to ~550MB/sec for block reads
  * block write latency decreases to a level roughly comparable to Ceph
  * block read latency increases to nearly 40ms
  * block rewrite latency accordingly gets absolutely terrible

Thoughts
---
* Better performance might be seen by using a RAID 0 device from onboard RAID or mdadm instead of an LVM JBOD, but that would increase the size of the failure domain and also make us dependent either on hardware RAID again or mdadm (and mdadm frustrates me).
* Ceph only has 3 nodes in this cluster, and so its performance numbers are likely to improve in larger clusters.
* While running the tests, the hypervisor showed unexpectedly high load average and iowait that would hover in the 10% range, occasionally spiking to 25%.
* Ephemeral volumes are apparently extremely sensitive to readahead settings, and so I will continue to test to try to find the optimal setting.